+++ src/main/java/io/github/ngspace/hudder/main/config/HudderConfig.java
@@ -43,7 +43,7 @@ public class HudderConfig {
 	@Expose public int xoffset = 1;
 	@Expose public int lineHeight = 10;
 	@Expose public int methodBuffer = 2;
-	@Expose public long backgroundcolor = 0x86353535;
+	@Expose public int backgroundcolor = 0x86353535;
 	@Expose public boolean background = true;
 	@Expose public boolean removegui = false;
 	@Expose public boolean limitrate = true;
